# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_delivery_http.DeviceResponse:
    properties:
      device_id:
        type: string
      id:
        type: string
      last_seen:
        type: string
      status:
        type: string
    type: object
  internal_delivery_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_delivery_http.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  internal_delivery_http.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Aimtec 2026 Laser Tag system — manages games, devices,
    players, teams and real-time events.
  title: Laser Tag API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      summary: Admin login
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout (invalidate token)
      tags:
      - auth
  /devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.DeviceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all devices
      tags:
      - devices
  /devices/available:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.DeviceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List available (online) devices
      tags:
      - devices
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (without quotes).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"